add comments

+++ utils.h
@@ -3,12 +3,12 @@
 
 struct node{
 	int id;
-	unsigned long int similarity;
+	int similarity;
 	struct node* next;
 };
 char** str_split( char* str, char delim, int* numSplits );
 struct node* readfile(char* filename, int* query, int dictionary_size, int* list_size);
 int* readquery(char* queryfile, int dictionary_size);
-void insertion_sort(unsigned long int *arr, int* ids, int length);
+void insertion_sort(int *arr, int* ids, int length);
 
 #endif
